+++ cli/ChangePasswordCommand.php
@@ -1,10 +1,12 @@
 <?php
+
 /**
  * @package    Grav\Plugin\Login
  *
  * @copyright  Copyright (C) 2014 - 2017 RocketTheme, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
+
 namespace Grav\Plugin\Console;
 
 use Grav\Common\User\Interfaces\UserCollectionInterface;