+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 3.0.0-rc.3
-## mm/dd/2019
+## 02/18/2019
 
 1. [](#new)
     * Requires Grav `1.6.0-rc.3`
@@ -7,7 +7,9 @@
     * Updated typehints from `User` to `UserInterface`
     * Use `$grav['users']` collection instead of deprecated static calls
     * Invalidate cache when modifying users from CLI
-    
+1. [](#bugfix)
+    * Fix login on registration (FlexUsers)   
+     
 # 3.0.0-rc.2
 ## 02/07/2019